First rooms and pills pooling
The first room and the two after it were changed to fit for the shield element. Lots of lasers and ghosts.
There is an object pooling for the pills in the GameControllerScript and a function that repositions the given amount of pills in a given position. The pills are located in a circle around that point.